+++ CHANGELOG.md
@@ -5,6 +5,12 @@ is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Note the the latest version is usually work in progress and may have not yet been released.
 
+# v0.0.7
+
+## Added
+
+- `--mask` flag to show commands to hide sensitive values
+
 # v0.0.6
 
 ## Added